Extract website recipe calls
